\name{ggspark-package}
\alias{ggspark-package}
\alias{ggspark}
\docType{package}
\title{\packageTitle{ggspark}}
\description{\packageDescription{ggspark}}
\section{Package Content}{\packageIndices{ggspark}}
\author{\packageAuthor{ggspark}}
\section{Maintainer}{\packageMaintainer{ggspark}}
\keyword{package}
